# docs/api-docs/master-data-swagger.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Master Data",
    "description" : "Personal Master Data API.",
    "version" : "0.1.0",
    "contact" : {
      "name" : "Adam Korzeniak",
      "email" : "korzeniak.adam@gmail.com",
      "url" : "https://www.adamkorzeniak.pl"
    }
  },
  "servers" : [ {
    "description" : "Stage/Test Server",
    "url" : "https://adamkorzeniak.pl:8443/v0"
  } ],
...
  "paths" : {
...
    "/Movie/genres" : {
      "get" : {
        "tags" : [ "Movie" ],
        "operationId" : "searchGenres",
        "summary" : "Searches genres",
        "description" : "Searches for genres using search parameters",
        "parameters" : [ {
          "in" : "query",
          "name" : "match-name",
          "description" : "Name of the genre",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : "Comedy"
          }
        }, {
          "in" : "query",
          "name" : "search-name",
          "description" : "Genre name has to contain that value",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : "drama"
          }
        }, {
          "in" : "query",
          "name" : "exist-name",
          "description" : "Genre name exists",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "example" : true
          }
        }, {
          "in" : "query",
          "name" : "order",
          "description" : "Order genres in default order using provided field",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : "name",
            "enum" : [ "name" ]
          }
        }, {
          "in" : "query",
          "name" : "order-asc",
          "description" : "Order genres in ascending order using provided field",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : "name",
            "enum" : [ "name" ]
          }
        }, {
          "in" : "query",
          "name" : "order-desc",
          "description" : "Order genres in descending order using provided field",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : "name",
            "enum" : [ "name" ]
          }
        } ],
        "responses" : {
          "200" : {
            "$ref" : "#/components/responses/genresDetails"
          },
          "400" : {
            "$ref" : "#/components/responses/badRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/unauthorized"
          },
          "404" : {
            "$ref" : "#/components/responses/notFound"
          },
          "405" : {
            "$ref" : "#/components/responses/methodNotAllowed"
          },
          "500" : {
            "$ref" : "#/components/responses/internalError"
          }
        }
      },
      "post" : {
        "tags" : [ "Movie" ],
        "operationId" : "addGenre",
        "summary" : "Adds a genre",
        "description" : "Creates a genre with provided details",
        "requestBody" : {
          "$ref" : "#/components/requestBodies/genreRequestDetails"
        },
        "responses" : {
          "201" : {
            "$ref" : "#/components/responses/genreDetails"
          },
          "400" : {
            "$ref" : "#/components/responses/badRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/unauthorized"
          },
          "404" : {
            "$ref" : "#/components/responses/notFound"
          },
          "405" : {
            "$ref" : "#/components/responses/methodNotAllowed"
          },
          "500" : {
            "$ref" : "#/components/responses/internalError"
          }
        }
      }
    },
    "/Movie/genres/{genreId}" : {
      "get" : {
        "tags" : [ "Movie" ],
        "operationId" : "getGenre",
        "summary" : "Retrieves a genre",
        "description" : "Retrieves a genre for given id",
        "parameters" : [ {
          "name" : "genreId",
          "in" : "path",
          "required" : true,
          "description" : "Return a genre with specific id",
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "example" : 42
          }
        } ],
        "responses" : {
          "200" : {
            "$ref" : "#/components/responses/genreDetails"
          },
          "400" : {
            "$ref" : "#/components/responses/badRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/unauthorized"
          },
          "404" : {
            "$ref" : "#/components/responses/notFound"
          },
          "405" : {
            "$ref" : "#/components/responses/methodNotAllowed"
          },
          "500" : {
            "$ref" : "#/components/responses/internalError"
          }
        }
      },
      "put" : {
        "tags" : [ "Movie" ],
        "operationId" : "updateGenre",
        "summary" : "Updates a genre",
        "description" : "Updates a genre with provided details",
        "parameters" : [ {
          "name" : "genreId",
          "in" : "path",
          "required" : true,
          "description" : "Return genre with specific id",
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "example" : 42
          }
        } ],
        "requestBody" : {
          "$ref" : "#/components/requestBodies/genreRequestDetails"
        },
        "responses" : {
          "200" : {
            "$ref" : "#/components/responses/genreDetails"
          },
          "400" : {
            "$ref" : "#/components/responses/badRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/unauthorized"
          },
          "404" : {
            "$ref" : "#/components/responses/notFound"
          },
          "405" : {
            "$ref" : "#/components/responses/methodNotAllowed"
          },
          "500" : {
            "$ref" : "#/components/responses/internalError"
          }
        }
      },
      "delete" : {
        "tags" : [ "Movie" ],
        "operationId" : "deleteGenre",
        "summary" : "Deletes a genre",
        "description" : "Deletes a genre with given id",
        "parameters" : [ {
          "name" : "genreId",
          "in" : "path",
          "required" : true,
          "description" : "genre id",
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "example" : 42
          }
        } ],
        "responses" : {
          "204" : {
            "$ref" : "#/components/responses/genreRemoved"
          },
          "400" : {
            "$ref" : "#/components/responses/badRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/unauthorized"
          },
          "404" : {
            "$ref" : "#/components/responses/notFound"
          },
          "405" : {
            "$ref" : "#/components/responses/methodNotAllowed"
          },
          "500" : {
            "$ref" : "#/components/responses/internalError"
          }
        }
      }
    },
    "/Error/errors" : {
      "get" : {
        "tags" : [ "Error" ],
        "operationId" : "searchErrors",
        "summary" : "Searches errors",
        "description" : "Search for errors using search parameters",
        "parameters" : [ {
          "in" : "query",
          "name" : "match-errorId",
          "description" : "errorId of the error",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : "error-1234"
          }
        }, {
          "in" : "query",
          "name" : "match-appId",
          "description" : "appId of the error",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : "app-1234"
          }
        }, {
          "in" : "query",
          "name" : "match-name",
          "description" : "name of the error",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : "Client issue"
          }
        }, {
          "in" : "query",
          "name" : "search-errorId",
          "description" : "error errorId has to contain that value",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : 1234
          }
        }, {
          "in" : "query",
          "name" : "search-appId",
          "description" : "error appId has to contain that value",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : 1234
          }
        }, {
          "in" : "query",
          "name" : "search-name",
          "description" : "error name has to contain that value",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : "client"
          }
        }, {
          "in" : "query",
          "name" : "search-details",
          "description" : "error details has to contain that value",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : "client"
          }
        }, {
          "in" : "query",
          "name" : "search-status",
          "description" : "error status has to contain that value",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : "client"
          }
        }, {
          "in" : "query",
          "name" : "search-location",
          "description" : "error location has to contain that value",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : "client"
          }
        }, {
          "in" : "query",
          "name" : "min-time",
          "description" : "error epoch time minimum value",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "example" : 1541111111
          }
        }, {
          "in" : "query",
          "name" : "max-time",
          "description" : "error epoch time maximum value",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "example" : 1541111111
          }
        }, {
          "in" : "query",
          "name" : "exist-details",
          "description" : "error details exists",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "example" : true
          }
        }, {
          "in" : "query",
          "name" : "order",
          "description" : "order errors in default order using provided field",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : "time",
            "enum" : [ "errorId", "appId", "name", "time" ]
          }
        }, {
          "in" : "query",
          "name" : "order-asc",
          "description" : "order errors in ascending order using provided field",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : "time",
            "enum" : [ "errorId", "appId", "name", "time" ]
          }
        }, {
          "in" : "query",
          "name" : "order-desc",
          "description" : "order errors in descending order using provided field",
          "required" : false,
          "schema" : {
            "type" : "string",
            "example" : "time",
            "enum" : [ "errorId", "appId", "name", "time" ]
          }
        } ],
        "responses" : {
          "200" : {
            "$ref" : "#/components/responses/errorsDetails"
          },
          "400" : {
            "$ref" : "#/components/responses/badRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/unauthorized"
          },
          "404" : {
            "$ref" : "#/components/responses/notFound"
          },
          "405" : {
            "$ref" : "#/components/responses/methodNotAllowed"
          },
          "500" : {
            "$ref" : "#/components/responses/internalError"
          }
        }
      },
      "post" : {
        "tags" : [ "Error" ],
        "operationId" : "addError",
        "summary" : "Adds an error",
        "description" : "Creates an error with provided details",
        "requestBody" : {
          "$ref" : "#/components/requestBodies/errorDetails"
        },
        "responses" : {
          "201" : {
            "$ref" : "#/components/responses/errorDetails"
          },
          "400" : {
            "$ref" : "#/components/responses/badRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/unauthorized"
          },
          "404" : {
            "$ref" : "#/components/responses/notFound"
          },
          "405" : {
            "$ref" : "#/components/responses/methodNotAllowed"
          },
          "500" : {
            "$ref" : "#/components/responses/internalError"
          }
        }
      }
    },
    "/Error/errors/{errorId}" : {
      "delete" : {
        "tags" : [ "Error" ],
        "operationId" : "deleteError",
        "summary" : "Deletes an error",
        "description" : "Deletes an error with given id",
        "parameters" : [ {
          "name" : "errorId",
          "in" : "path",
          "required" : true,
          "description" : "error id",
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "example" : 42
          }
        } ],
        "responses" : {
          "204" : {
            "$ref" : "#/components/responses/errorRemoved"
          },
          "400" : {
            "$ref" : "#/components/responses/badRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/unauthorized"
          },
          "404" : {
            "$ref" : "#/components/responses/notFound"
          },
          "405" : {
            "$ref" : "#/components/responses/methodNotAllowed"
          },
          "500" : {
            "$ref" : "#/components/responses/internalError"
          }
        }
      }
    }
  },
  "security" : [ {
    "basicAuth" : [ ]
  } ],
  "components" : {
    "securitySchemes" : {
      "basicAuth" : {
        "type" : "http",
        "scheme" : "basic",
        "description" : "For Stage environment use below credentials:<br>Username: <b>test</b><br>Password: <b>test123</b><br><br>"
      }
    },
    "schemas" : {
...
      "Genre" : {
        "type" : "object",
        "required" : [ "name" ],
        "properties" : {
          "id" : {
            "description" : "Genre id",
            "type" : "integer",
            "format" : "int64",
            "minimum" : 1,
            "example" : 789,
            "readOnly" : true
          },
          "name" : {
            "description" : "Genre name",
            "type" : "string",
            "minLength" : 1,
            "example" : "Drama"
          }
        }
      },
      "GenreList" : {
        "type" : "array",
        "items" : {
          "$ref" : "#/components/schemas/Genre"
        }
      },
      "Error" : {
        "type" : "object",
        "required" : [ "id", "errorCode", "appId" ],
        "properties" : {
          "id" : {
            "description" : "Error id in database",
            "type" : "integer",
            "format" : "int64",
            "example" : 987
          },
          "errorCode" : {
            "description" : "Error code",
            "type" : "string",
            "example" : "error-1234"
          },
          "appId" : {
            "type" : "string",
            "example" : "app-1234"
          },
          "name" : {
            "type" : "string",
            "example" : "client issue"
          },
          "details" : {
            "type" : "string",
            "example" : "unexpected client error"
          },
          "status" : {
            "type" : "string",
            "example" : "status"
          },
          "url" : {
            "type" : "string",
            "example" : "Movies/-1"
          },
          "stack" : {
            "type" : "string",
            "example" : "stack trace"
          },
          "time" : {
            "type" : "integer",
            "format" : "int64",
            "example" : 1541111111
          }
        }
      },
      "ErrorList" : {
        "type" : "array",
        "items" : {
          "$ref" : "#/components/schemas/Error"
        }
      },
      "ErrorResponse" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "description" : "error code",
            "type" : "string",
            "example" : "REQ001"
          },
          "title" : {
            "description" : "error title",
            "type" : "string",
            "example" : "BAD REQUEST"
          },
          "message" : {
            "description" : "error message",
            "type" : "string",
            "example" : "Missing field title"
          }
        }
      }
    },
    "examples" : {
      "BadRequestErrorResponse" : {
        "value" : {
          "code" : "BAD_REQUEST",
          "title" : "Bad Request",
          "message" : "Bad Request"
        }
      },
      "NotAuthorizedErrorResponse" : {
        "value" : {
          "code" : "NOT_AUTHORIZED",
          "title" : "Not Authorized",
          "message" : "Not Authorized"
        }
      },
      "NotFoundErrorResponse" : {
        "value" : {
          "code" : "NOT_Found",
          "title" : "Not Found",
          "message" : "Not Found"
        }
      },
      "MethodNotAllowedErrorResponse" : {
        "value" : {
          "code" : "METHOD_NOT_ALLOWED",
          "title" : "Not Found",
          "message" : "Not Found"
        }
      },
...
      "InternalErrorResponse" : {
        "value" : {
          "code" : "INTERNAL_ERROR",
          "title" : "Not Found",
          "message" : "Not Found"
        }
      }
    },
    "requestBodies" : {
...
      "genreRequestDetails" : {
        "description" : "Genre details",
        "required" : true,
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Genre"
            }
          }
        }
      },
      "errorDetails" : {
        "description" : "Error details",
        "required" : true,
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "responses" : {
...
      "genreDetails" : {
        "description" : "Genre details returned",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Genre"
            }
          }
        }
      },
      "genresDetails" : {
        "description" : "List of genres returned",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/GenreList"
            }
          }
        }
      },
      "genreRemoved" : {
        "description" : "Genre removed"
      },
      "errorsDetails" : {
        "description" : "List of errors returned",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorList"
            }
          }
        }
      },
      "errorDetails" : {
        "description" : "Created error details returned",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Error"
            }
          }
        }
      },
      "errorRemoved" : {
        "description" : "Error removed"
      },
...
      "badRequest" : {
        "description" : "The request could not be understood by the server due to malformed syntax",
        "content" : {
          "application/json" : {
            "example" : {
              "$ref" : "#/components/examples/BadRequestErrorResponse"
            },
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "unauthorized" : {
        "description" : "User is not authorised",
        "content" : {
          "application/json" : {
            "example" : {
              "$ref" : "#/components/examples/NotAuthorizedErrorResponse"
            },
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "notFound" : {
        "description" : "Server could not found requested resource",
        "content" : {
          "application/json" : {
            "example" : {
              "$ref" : "#/components/examples/NotFoundErrorResponse"
            },
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "methodNotAllowed" : {
        "description" : "Method is not supported for given resource",
        "content" : {
          "application/json" : {
            "example" : {
              "$ref" : "#/components/examples/MethodNotAllowedErrorResponse"
            },
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "internalError" : {
        "description" : "The server encountered an unexpected condition which prevented it from fulfilling the request.",
        "content" : {
          "application/json" : {
            "example" : {
              "$ref" : "#/components/examples/InternalErrorResponse"
            },
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}